test_runner: rerun all failed tests
